requirements.txt
boto3==1.4.4
ruamel.yaml==0.13.14
